create api controller, move some routes
