+++ app/Http/Kernel.php
@@ -31,15 +31,16 @@ class Kernel extends HttpKernel
         'auth.basic' => \Illuminate\Auth\Middleware\AuthenticateWithBasicAuth::class,
         'guest' => \App\Http\Middleware\RedirectIfAuthenticated::class,
         'member' => \App\Http\Middleware\RedirectIfNotAMember::class,
+        'participant' => \App\Http\Middleware\RedirectIfNotAParticipant::class,
         'cors' => \App\Http\Middleware\CORS::class,
         'bindings' => \Illuminate\Routing\Middleware\SubstituteBindings::class,
         'can' => \Illuminate\Auth\Middleware\Authorize::class,
         'privacy' => \App\Http\Middleware\CheckPrivacy::class,
+        'throttle' => \Illuminate\Routing\Middleware\ThrottleRequests::class,
     ];
 
     protected $middlewareGroups = [
         'web' => [
-
             \App\Http\Middleware\EncryptCookies::class,
             \Illuminate\Cookie\Middleware\AddQueuedCookiesToResponse::class,
             \Illuminate\Session\Middleware\StartSession::class,
@@ -51,7 +52,8 @@ class Kernel extends HttpKernel
 
         'api' => [
             'throttle:60,1',
-            'auth:api',
+            'cors',
+            // 'auth:api',
         ],
     ];
 }
