+++ routes/web.auth.php
@@ -2,8 +2,6 @@
 
 declare(strict_types=1);
 
-use App\Http\Middleware\RedirectIfNotAMember;
-use App\Http\Middleware\RedirectIfNotAParticipant;
 use Illuminate\Support\Facades\Route;
 
 Route::get('home', 'PagesController@home');
@@ -19,7 +17,7 @@ Route::put('profile/password', 'ProfileController@passwordSave');
 Route::get('profile/on-camp', 'ProfileController@onCamp');
 Route::put('profile/on-camp', 'ProfileController@onCampSave');
 
-Route::middleware([RedirectIfNotAMember::class])->group(function () {
+Route::middleware('member')->group(function () {
     Route::get('profile/add-course', 'ProfileController@addCourse');
     Route::put('profile/add-course', 'ProfileController@addCourseSave');
     Route::get('profile/edit-course/{course}', 'ProfileController@editCourse');
@@ -32,7 +30,7 @@ Route::middleware([RedirectIfNotAMember::class])->group(function () {
     Route::get('profile/reviews/{event}', 'ProfileController@reviews');
 });
 
-Route::middleware([RedirectIfNotAParticipant::class])->group(function () {
+Route::middleware('participant')->group(function () {
     Route::get('profile/edit-camp/{event}', 'ProfileController@editCamp');
     Route::put('profile/edit-camp/{event}', 'ProfileController@editCampSave');
 });
@@ -150,7 +148,7 @@ Route::delete(
 );
 Route::get('members/{member}/delete', 'MembersController@delete');
 
-Route::middleware(RedirectIfNotAMember::class)->group(function () {
+Route::middleware('member')->group(function () {
     Route::get('lists', 'ListsController');
     Route::get('graphs', 'GraphsController');
 });
