+++ routes/web.public.php
@@ -5,15 +5,8 @@ declare(strict_types=1);
 use Illuminate\Support\Facades\Route;
 use Updater\Http\Controllers\UpdateController;
 
-// iDeal integration
-Route::post('iDeal-webhook', 'iDealController@webhook');
-Route::get('iDeal-response/{participant}/{event}', 'iDealController@response');
-
 // Pages
 Route::get('privacy', 'PagesController@showPrivacyStatement');
-Route::get('cal/{type}', 'PagesController@cal')->middleware('cors');
-Route::get('camp-info/{camp}', 'PagesController@campInfo');
-Route::get('camps-report', 'PagesController@campsReport');
 
 // Review things
 Route::get('enquete/{event}', 'ReviewsController@review');
@@ -24,5 +17,3 @@ Route::get('events/icalendar', 'EventsController@iCalendar');
 
 Route::get('updater/update', [UpdateController::class, 'update']);
 Route::get('updater/version', [UpdateController::class, 'version']);
-
-Route::post('api/contact-form', 'ContactFormController@send')->middleware('cors');
